[Dummy Commit] Preserve parallel loop analysis starting code.
This is a dummy commit to try another direction. So far, I have been trying
to analyse parallel loops to extract needed information such as the type of
the iteration variable, iteration count, etc. This is guided by the definition
of a canonical loop as specified by OpenMP. I will instead canonicalize loops
and, therefore, won't need any pieces of info other than the iteration count.